feat: Add ESLint configuration and update package.json
This commit adds a new `.eslintrc.json` file to the `myshop` directory, which includes the ESLint configuration for the project. It sets the environment to support browser, ES2021, and Node.js, extends the `airbnb-base` and `plugin:vue/vue3-essential` configurations, and includes the necessary plugin dependencies. The ESLint rules are left empty for now.

In addition, the `package.json` file in the `myshop` directory is updated. The following dependencies are added: `eslint`, `eslint-config-airbnb-base`, `eslint-plugin-import`, `eslint-plugin-vue`, and `vite-plugin-eslint`. The version of `vue` is also updated to `3.3.13`. The `devDependencies` section is updated to remove the `vite` dependency, as it is already included in the dependencies section. The versions of the added dependencies are specified.

There are also minor changes to the `index.html` file, where the path to the `main.js` file is updated to use a relative path. Finally, the `pnpm-lock.yaml` file includes updates to the versions of the added dependencies.