fix: Handle API response errors and throw an error message
The code in `ProductItem.vue` and `HomeView.vue` was modified to handle API response errors. Previously, the errors were only logged to the console, but now they are thrown as an error message. This change ensures that any error encountered during the API call is properly handled and displayed to the user.