Update router title and description handling
The router logic in `myshop/src/router/index.js` has been modified to handle the page title and description. Previously, the page title was set to the value of `to.name` or 'MyShop', and the page description was set to 'Description par défaut de la page.' Now, if `to.meta.title` is defined, the page title will be set to that value instead. This change ensures that the correct title and description are displayed for each page in MyShop.