Update image upload headers with authentication token
This commit modifies the `ProductsEdit.vue` file in the `myshop/src/views/admin` directory. The `uploadImage` function now includes an authentication token in the headers when making the `POST` request to the image API. The token is obtained by calling the `getAuthenticationToken` function. This change ensures that the image upload request is authenticated.